# tsoa/swagger.json
{
	"components": {
...
		"schemas": {
			"IPayment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"cardNumber": {
						"type": "string"
					},
					"userID": {
						"type": "string"
					},
					"cvv": {
						"type": "string"
					},
					"placeHolderName": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					},
					"expirationDate": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"cardNumber",
					"cvv",
					"placeHolderName",
					"receiver",
					"expirationDate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/payment/buy-course/{paymentId}/{courseId}": {
			"get": {
				"operationId": "BuyCourse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "paymentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "courseId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/payment/create": {
			"post": {
				"operationId": "CreatePayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPayment"
							}
						}
					}
				}
			}
		},
		"/api/payment/update/{id}": {
			"put": {
				"operationId": "UpdatePayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPayment"
							}
						}
					}
				}
			}
		},
		"/api/payment/delete/{id}": {
			"delete": {
				"operationId": "DeletePayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}